INCOMPATIBLE: move main module
The main cluster modules have been moved from the root
of the project to a regular module (instance) in the
modules folder.
This is incompatible. A state migration cannot be defined
because terraform cannot handle the move of template
resources. This type of resource saves absolute path names
in the state (never a good idea). These are not valid
anymore, if the resource is moved.